src/main/webapp/app/entities/tipo-denuncia/delete/tipo-denuncia-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { TipoDenunciaService } from '../service/tipo-denuncia.service';

import { TipoDenunciaDeleteDialogComponent } from './tipo-denuncia-delete-dialog.component';

describe('TipoDenuncia Management Delete Component', () => {
  let comp: TipoDenunciaDeleteDialogComponent;
  let fixture: ComponentFixture<TipoDenunciaDeleteDialogComponent>;
  let service: TipoDenunciaService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [TipoDenunciaDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(TipoDenunciaDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(TipoDenunciaDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(TipoDenunciaService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
